EditEnd/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2
asyncio-atexit==1.0.1
attrs==24.2.0
bce-python-sdk==0.9.23
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorlog==6.8.2
dashscope==1.20.11
distro==1.9.0
erniebot==0.5.9
erniebot-agent==0.5.2
Flask==3.0.3
Flask-Cors==5.0.0
frozenlist==1.4.1
future==1.0.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-path==0.3.3
jsonschema-specifications==2023.12.1
langchain==0.3.3
langchain-core==0.3.10
langchain-text-splitters==0.3.0
langsmith==0.1.134
lazy-object-proxy==1.10.0
MarkupSafe==3.0.1
multidict==6.1.0
mysql-connector==2.2.9
numpy==1.26.4
openai==1.51.2
openapi-schema-validator==0.6.2
openapi-spec-validator==0.7.1
orjson==3.10.7
packaging==24.1
pathable==0.4.3
propcache==0.2.0
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rpds-py==0.20.0
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
tenacity==8.5.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
Werkzeug==3.0.4
wheel==0.44.0
yarl==1.15.2
